% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputValidation.R
\name{checkInputBounds}
\alias{checkInputBounds}
\title{checkInputBounds}
\usage{
checkInputBounds(input)
}
\arguments{
\item{input}{GUi Inputs}
}
\value{
T for correct user input, F for a faulty configuration
}
\description{
Checks whether or not the user inserted a faulty configuration for the dimension/bounds
}
