% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputValidation.R
\name{checkInputObjectiveFunction}
\alias{checkInputObjectiveFunction}
\title{checkInputObjectiveFunction}
\usage{
checkInputObjectiveFunction(input)
}
\arguments{
\item{input}{GUI Inputs}
}
\value{
T for correct user input, F for a faulty configuration
}
\description{
Checks whether or not the user inserted a faulty configuration for an
objective Function from the R-Environment
}
