% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Runs on package attach}
\usage{
.onAttach(libname, pkgname)
}
\value{
None
}
\description{
The package startup message is defined here. It shall give a clue on how to
start the owosGUI and which suggests might want to be installed.
}
\keyword{internal}
