% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputCSVColumns.R
\name{getAmountCsvCheckBoxes}
\alias{getAmountCsvCheckBoxes}
\title{Choosing the input columns}
\usage{
getAmountCsvCheckBoxes()
}
\value{
number of columns in the packageData Environment
}
\description{
Choosing the input columns
}
\keyword{internal}
