% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiGetters.R
\name{getClosureVariableFromUI}
\alias{getClosureVariableFromUI}
\title{getClosureVariableFromUI}
\usage{
getClosureVariableFromUI(input, inputName, asText = F)
}
\arguments{
\item{input}{shiny server input}

\item{inputName}{missing docu}

\item{asText}{for log generation set 'asText' = T}
}
\value{
closureVar
}
\description{
Missing Docu
}
\keyword{internal}
