% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiGetters.R
\name{getControlList}
\alias{getControlList}
\title{getControlList}
\usage{
getControlList(input, strID, asText = FALSE)
}
\arguments{
\item{input}{shiny server input}

\item{strID}{group name in xml file}

\item{asText}{for log generation set 'asText' = TRUE}
}
\value{
controlList
}
\description{
Load a control list based on user input in the xml generated configurators.
}
\keyword{internal}
