% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDataStorage.R
\name{getEnvData}
\alias{getEnvData}
\title{Access a variable in the packageData Environment}
\usage{
getEnvData(strName)
}
\arguments{
\item{strName}{String/Character vector. Name of variable that is to be accessed.}
}
\value{
current value of the variable
}
\description{
Access a variable in the packageData Environment
}
\keyword{internal}
