% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiGetters.R
\name{getExtraParametersList}
\alias{getExtraParametersList}
\title{getExtraParametersList}
\usage{
getExtraParametersList(input, strID, asText = F)
}
\arguments{
\item{input}{shiny server input}

\item{strID}{group name in xml file}

\item{asText}{for log generation set 'asText' = T}
}
\value{
extraValue
}
\description{
Get value of extra field below normal setup in gui
}
\keyword{internal}
