% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUi.R
\name{getUIPage}
\alias{getUIPage}
\title{Defines the shiny UI-part of the GUI}
\usage{
getUIPage()
}
\value{
None
}
\description{
Generates the UI part of the GUI.
This method is used internally in the starting process,
manual use of this function is not advised.
}
