% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiGetters.R
\name{getXMLVarListUI}
\alias{getXMLVarListUI}
\title{getXMLVarListUI}
\usage{
getXMLVarListUI(input, rootName, varList, asText = FALSE)
}
\arguments{
\item{input}{shiny server input}

\item{rootName}{xml root name group name in xml file}

\item{varList}{the variable name to get}

\item{asText}{for log generation set 'asText' = TRUE}
}
\value{
varList
}
\description{
Load values of a variablelist from ui based on its xml name
}
\keyword{internal}
