% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDataStorage.R
\docType{data}
\name{packageDataEnv}
\alias{packageDataEnv}
\title{Package Data Environment}
\format{An object of class \code{environment} of length 0.}
\usage{
packageDataEnv
}
\description{
Data Environment for storage and variable exchange between functions
}
\keyword{datasets}
