% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordNewTest.R
\name{recordNewTest}
\alias{recordNewTest}
\title{Record a new GUI test sequence}
\usage{
recordNewTest()
}
\value{
None
}
\description{
Simply run 'recordNewTest()' and the test creation GUI will pop up.
}
\keyword{internal}
