% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{runGUI}
\usage{
runGUI()
}
\value{
None
}
\description{
Run the starting command of the EventDetectGUI. Opens the graphical shiny application through which the user
can acess the algorithms and visualizations.
}
