% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDataStorage.R
\name{setEnvData}
\alias{setEnvData}
\title{Set a value to a variable in the packageData Environment}
\usage{
setEnvData(strName, value)
}
\arguments{
\item{strName}{String/Character vector. Name of variable that is to be accessed.}

\item{value}{Any value that shall be set to the respective variable}
}
\value{
None
}
\description{
Set a value to a variable in the packageData Environment
}
\keyword{internal}
