% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlDynamicUI.R
\name{xmlGetAllConfiguredControlElements}
\alias{xmlGetAllConfiguredControlElements}
\title{xmlGetAllConfiguredControlElements}
\usage{
xmlGetAllConfiguredControlElements()
}
\value{
List with all configured optimizers, designGenerators and modelBuilders,
as well as their respective parameters.
}
\description{
Parses the spotconfig xml file. Reads all entries for optimizers, designGenerators and modelBuilders.
For each configured entrie, a list with the entries name and all of its configurable parameters is returned.
}
\examples{
spotGUI:::xmlGetAllConfiguredControlElements()
}
