% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlDynamicUI.R
\name{xmlGetRootElement}
\alias{xmlGetRootElement}
\title{xmlGetAllConfiguredControlElements}
\usage{
xmlGetRootElement()
}
\value{
List with all configured preProcess, postProcess and algorithms
as well as their respective parameters.
}
\description{
Parses the config.xml file. Reads all entries for preProcess, postProcess and algorithm
For each configured entrie, a list with the entries name and all of its configurable parameters is returned.
}
