% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServer.R
\name{getServer}
\alias{getServer}
\title{Generate Server Part of GUI}
\usage{
getServer(input, output, session)
}
\arguments{
\item{input}{shiny UI-input}

\item{output}{shiny UI-output}

\item{session}{shiny UI-session}
}
\value{
None
}
\description{
Generates the server part of the GUI.
This method is used internally in the starting process of the GUI.
Manual use of this function is not advised.
}
