% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiGetters.R
\name{getXMLVariableUI}
\alias{getXMLVariableUI}
\title{getXMLVariableUI}
\usage{
getXMLVariableUI(input, rootName, var, asText = FALSE)
}
\arguments{
\item{input}{shiny server input}

\item{rootName}{xml root name group name in xml file}

\item{var}{the variable name to get}

\item{asText}{for log generation set 'asText' = TRUE}
}
\value{
var
}
\description{
Load values of a variablelist from ui based on its xml name
}
\keyword{internal}
